# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EcoLoop",
        "description": "Application de dons et d'échanges d'objets durables",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/articles": {
            "get": {
                "tags": [
                    "Article"
                ],
                "summary": "liste de tous les categories",
                "operationId": "8b1e406f12c2ca9a811359ffde9f181d",
                "responses": {
                    "200": {
                        "description": "succes"
                    }
                }
            },
            "post": {
                "tags": [
                    "Article"
                ],
                "summary": "Ajout d'un article",
                "operationId": "59d8c76516c6503a609d3ca85de30c47",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "libelle": {
                                        "type": "string"
                                    },
                                    "image": {
                                        "description": "Fichier de photo",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "contenu": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Article ajouté avec succées"
                    },
                    "401": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/nombresArticles": {
            "get": {
                "tags": [
                    "Article"
                ],
                "summary": "Nombres d'articles",
                "operationId": "dfad6fc5e839eac53bd823cc0d89ccfd",
                "responses": {
                    "200": {
                        "description": "succes"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/articles/{article}": {
            "get": {
                "tags": [
                    "Article"
                ],
                "summary": "Details d'un article",
                "operationId": "ae2592b01411f61e964d74c85da6a589",
                "parameters": [
                    {
                        "name": "article",
                        "in": "path",
                        "description": "ID de l'article à afficher",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "libelle": {
                                            "type": "string"
                                        },
                                        "image": {
                                            "type": "string",
                                            "example": "path/to/photo.jpg"
                                        },
                                        "contenu": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                }
            },
            "post": {
                "tags": [
                    "Article"
                ],
                "summary": "Modification d'un article",
                "operationId": "2177d924627764d948b15fe8695a694c",
                "parameters": [
                    {
                        "name": "article",
                        "in": "path",
                        "description": "ID de l'article à modifier",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "libelle": {
                                        "type": "string"
                                    },
                                    "image": {
                                        "description": "Fichier de photo",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "contenu": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Article modifié avec succées"
                    },
                    "401": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Article"
                ],
                "summary": "Supprimer un article",
                "operationId": "858d9cb40f8edbad7416e5fb71ab8248",
                "parameters": [
                    {
                        "name": "article",
                        "in": "path",
                        "description": "ID de l'article à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "succes"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}